Corrected topic class moving in all classes
