--- src/main/java/model/datatypes/Device.java
@@ -1,7 +1,0 @@
-package model.datatypes;
-
-/**
- * Created by Gorka Olalde on 2/6/16.
- */
-public class Device {
-}
